requirements.txt
bleak
dbus-fast
PySDL2
pyinstaller
